Small changes
Material changes, surroundings, bug fixes and selection indicator
