+++ Assets/Materials/Black02.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a5c69e5c8fe59da4da1b6dc92bbe365d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
