+++ Assets/Materials/Selected.mat.meta
@@ -1,7 +1,8 @@
 fileFormatVersion: 2
-guid: f3968442c09f9f6458b6f7672879e8de
-DefaultImporter:
+guid: 4576ef9caedbc434cb1adad036bb941f
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
