+++ Assets/Materials/White02.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8e162f589db926f4e96004dc82a60e9d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
