+++ Assets/Materials/Wood.mat.meta
@@ -1,7 +1,8 @@
 fileFormatVersion: 2
-guid: 72434c6a85e918c459132171574b1612
-DefaultImporter:
+guid: 7cf81a053c3249e49a0d2f206c1d6e05
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
