--- Assets/Models/Bishop.mtl
@@ -1,10 +1,0 @@
-# Blender MTL File: 'Bishop.blend'
-# Material Count: 1
-
-newmtl None
-Ns 500
-Ka 0.8 0.8 0.8
-Kd 0.8 0.8 0.8
-Ks 0.8 0.8 0.8
-d 1
-illum 2
