+++ Assets/Prefabs/Bishop.prefab
@@ -58,7 +58,7 @@ MeshRenderer:
   m_RenderingLayerMask: 1
   m_RendererPriority: 0
   m_Materials:
-  - {fileID: 7741625113677808684, guid: ac8d9329685b1ce4aa01115fa36a25ec, type: 3}
+  - {fileID: 2100000, guid: f3fac794f134da340862d4290b97bb36, type: 2}
   m_StaticBatchInfo:
     firstSubMesh: 0
     subMeshCount: 0
@@ -89,7 +89,7 @@ GameObject:
   m_Component:
   - component: {fileID: 6557694717295152259}
   - component: {fileID: 3888570938742539713}
-  - component: {fileID: -8420778351005145511}
+  - component: {fileID: 6406787729958804540}
   m_Layer: 0
   m_Name: Bishop
   m_TagString: Untagged
@@ -125,7 +125,7 @@ BoxCollider:
   serializedVersion: 2
   m_Size: {x: 2, y: 5, z: 2}
   m_Center: {x: 0, y: 2.5, z: 0}
---- !u!114 &-8420778351005145511
+--- !u!114 &6406787729958804540
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -134,6 +134,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 5785520463763805753}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 73b17891006c6fd4b99a6e2cd70ec06f, type: 3}
+  m_Script: {fileID: 11500000, guid: 554b25ff33220e74e848cb3dea90aa6c, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  white: 1
