+++ Assets/Prefabs/Queen_White.prefab
@@ -58,7 +58,7 @@ MeshRenderer:
   m_RenderingLayerMask: 1
   m_RendererPriority: 0
   m_Materials:
-  - {fileID: 7741625113677808684, guid: 89f6caccfaa9f734a82cb3876516b319, type: 3}
+  - {fileID: 2100000, guid: f3fac794f134da340862d4290b97bb36, type: 2}
   m_StaticBatchInfo:
     firstSubMesh: 0
     subMeshCount: 0
@@ -89,7 +89,7 @@ GameObject:
   m_Component:
   - component: {fileID: 6960122547706602716}
   - component: {fileID: 6798486899992031804}
-  - component: {fileID: 6899142699326500017}
+  - component: {fileID: 3028760530841770652}
   m_Layer: 0
   m_Name: Queen_White
   m_TagString: Untagged
@@ -104,7 +104,7 @@ Transform:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 7772230234695106150}
-  m_LocalRotation: {x: 0, y: -0, z: -0, w: 1}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 0.5, z: 4}
   m_LocalScale: {x: 0.4, y: 0.4, z: 0.4}
   m_Children:
@@ -125,7 +125,7 @@ BoxCollider:
   serializedVersion: 2
   m_Size: {x: 2.5, y: 5, z: 2.5}
   m_Center: {x: 0, y: 2.5, z: 0}
---- !u!114 &6899142699326500017
+--- !u!114 &3028760530841770652
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -134,7 +134,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 7772230234695106150}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 73b17891006c6fd4b99a6e2cd70ec06f, type: 3}
+  m_Script: {fileID: 11500000, guid: 58540406eb15be34295982bfa2e84939, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   white: 0
