+++ Assets/Scripts/Pieces/Types.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b061a020995458d4488aade44026dcf3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
