+++ Assets/Scripts/Pieces/Types/Bishop.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 554b25ff33220e74e848cb3dea90aa6c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
