+++ Assets/Scripts/Pieces/Types/King.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 49a1baa713aa5e849955fc112b213050
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
