+++ Assets/Scripts/Pieces/Types/Knight.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3ac46f7722ff90a43b7c2efd0a5812e8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
