+++ Assets/Scripts/Pieces/Types/Queen.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 58540406eb15be34295982bfa2e84939
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
